bug fix in the puzzle solver and progress on puzzle generator
Done: handle the case where the current symbol is not an empty symbol
and it has the same value as s. In this case update_symbol should return
false without registering an error condition. If I do that then
example4
fails on the backtracking and returns unsolvable configuration
error on the 11317th iteration which means that another bug is lurking
around